Feat : read 구현
